requirements.txt
# torch==2.4.0
# torchvision==0.19.0
gradio==6.1.0
gradio_client==2.0.1
trimesh==4.10.1
matplotlib>=3.7.0
easydict>=1.13
jaxtyping>=0.3.3
plyfile>=1.1.3
videoio>=0.3.0
numpy>=1.24.0,<2
scipy>=1.10.0
Pillow>=10.0.0
tqdm>=4.66.0
PyYAML>=6.0.0
ninja>=1.11
huggingface-hub>=0.24.0

# use ffmpeg backend for writing mp4 with imageio.mimwrite
imageio[ffmpeg]>=2.31.0
imageio-ffmpeg>=0.5.0

einops>=0.7.0
# xformers==0.0.27.post2
xformers==0.0.35
opencv-python==4.11.0.86

# gsplat
git+https://github.com/QitaoZhao/gsplat.git